Split functionality into seperate classes, de-bloating the main CS file
